+++ Rest/src/main/java/controller/MonthlyBudgetController.java
@@ -1,6 +1,7 @@
 package controller;
 
 
+import domain.User;
 import dto.ExpenseDto;
 
 import dto.MonthlyBudgetDto;
@@ -12,6 +13,8 @@ import service.IService;
 import service.ServiceEmptyResponse;
 import service.exception.ServiceException;
 
+import java.util.Optional;
+
 @CrossOrigin
 @RestController
 @RequestMapping("api/expense-management")
@@ -20,20 +23,26 @@ public class MonthlyBudgetController {
     @Autowired
     private IService service;
 
-    @DeleteMapping("/delete-monthly-budget/{budgetId}/{userId}")
+    @DeleteMapping("/delete-monthly-budget/{budgetId}")
     public ResponseEntity<?> deleteMonthlyBudget(
             @PathVariable int budgetId,
-            @PathVariable int userId
+            @RequestBody String body
     ){
-        // TODO - get userId using authorization method; return unauthorized if is not valid; remove userId from path
-        ServiceEmptyResponse response = service.deleteMonthlyBudget(budgetId,userId);
-        switch (response.getStatus()){
-            case 403:
-                return new ResponseEntity<>(response.getErrorMessage(),HttpStatus.FORBIDDEN);
-            case 500:
-                return new ResponseEntity<>(response.getErrorMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        Optional<User> user=service.getTokenUser(body);
+        if(user.isPresent()) {
+            int userId = user.get().getId();
+            ServiceEmptyResponse response = service.deleteMonthlyBudget(budgetId, userId);
+            switch (response.getStatus()) {
+                case 403:
+                    return new ResponseEntity<>(response.getErrorMessage(), HttpStatus.FORBIDDEN);
+                case 500:
+                    return new ResponseEntity<>(response.getErrorMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+            return new ResponseEntity<>(HttpStatus.OK);
         }
-        return new ResponseEntity<>(HttpStatus.OK);
+        else{
+            return new ResponseEntity<>("Unauthorized",HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PostMapping("add-monthly-budget")
